Add method #PromptPlayerMove()
This UI method should be able to take an optional array of integers to
restrict which inputs are acceptable. If not included, any integer input
is valid. Input should be converted to an integer and returned.

If input is invalidated, the message of the prompt should repeat and
input should be received again. The process repeats until integer and
"filter" conditions are met.